HOSTDEV-976 jissue commit improvements
This commit resolves both HOSTDEV-976 and HOSTDEV-954 as:
* it injects the issue summary and description into the commit message
* opens the GIT_EDITOR for editing if no message was provided in the command-line arguments

Summary:
automatically add to commit message the issue summary + part of the description

Description:
When doing 'git log' on izbox mainline I've noticed several commits that had only the issue number in their description.

Obviously we have to assign a ticket to a commit, but putting only the issue number makes git log useless... Therefore I propose that jissue should append to a commit message (that should be allowed to be empty from now on) the issue's subject and a truncated version of the issue's description (X lines or X characters, etc.).

This will make writing a commit message unnecessary since all the information will be in the issue (a good thing).